Add download package to jetpack.backup
