+++ WordPress/src/main/java/org/wordpress/android/modules/AppComponent.java
@@ -45,10 +45,10 @@ import org.wordpress.android.ui.domains.DomainSuggestionsFragment;
 import org.wordpress.android.ui.gif.GifPickerActivity;
 import org.wordpress.android.ui.history.HistoryAdapter;
 import org.wordpress.android.ui.history.HistoryDetailContainerFragment;
-import org.wordpress.android.ui.jetpack.backup.BackupDownloadActivity;
-import org.wordpress.android.ui.jetpack.backup.complete.BackupDownloadCompleteFragment;
-import org.wordpress.android.ui.jetpack.backup.details.BackupDownloadDetailsFragment;
-import org.wordpress.android.ui.jetpack.backup.progress.BackupDownloadProgressFragment;
+import org.wordpress.android.ui.jetpack.backup.download.BackupDownloadActivity;
+import org.wordpress.android.ui.jetpack.backup.download.complete.BackupDownloadCompleteFragment;
+import org.wordpress.android.ui.jetpack.backup.download.details.BackupDownloadDetailsFragment;
+import org.wordpress.android.ui.jetpack.backup.download.progress.BackupDownloadProgressFragment;
 import org.wordpress.android.ui.jetpack.scan.ScanActivity;
 import org.wordpress.android.ui.jetpack.scan.ScanFragment;
 import org.wordpress.android.ui.main.AddContentAdapter;
