+++ WordPress/src/main/java/org/wordpress/android/modules/ApplicationModule.java
@@ -15,8 +15,8 @@ import org.wordpress.android.ui.accounts.signup.SettingsUsernameChangerFragment;
 import org.wordpress.android.ui.accounts.signup.UsernameChangerFullScreenDialogFragment;
 import org.wordpress.android.ui.domains.DomainRegistrationDetailsFragment.CountryPickerDialogFragment;
 import org.wordpress.android.ui.domains.DomainRegistrationDetailsFragment.StatePickerDialogFragment;
-import org.wordpress.android.ui.jetpack.backup.BackupDownloadStep;
-import org.wordpress.android.ui.jetpack.backup.BackupDownloadStepsProvider;
+import org.wordpress.android.ui.jetpack.backup.download.BackupDownloadStep;
+import org.wordpress.android.ui.jetpack.backup.download.BackupDownloadStepsProvider;
 import org.wordpress.android.ui.mediapicker.loader.TenorGifClient;
 import org.wordpress.android.ui.posts.BasicDialog;
 import org.wordpress.android.ui.reader.ReaderPostWebViewCachingFragment;
