+++ WordPress/src/main/java/org/wordpress/android/modules/ViewModelModule.java
@@ -6,10 +6,10 @@ import androidx.lifecycle.ViewModelProvider;
 import org.wordpress.android.ui.JetpackRemoteInstallViewModel;
 import org.wordpress.android.ui.activitylog.list.filter.ActivityLogTypeFilterViewModel;
 import org.wordpress.android.ui.domains.DomainRegistrationMainViewModel;
-import org.wordpress.android.ui.jetpack.backup.BackupDownloadViewModel;
-import org.wordpress.android.ui.jetpack.backup.complete.BackupDownloadCompleteViewModel;
-import org.wordpress.android.ui.jetpack.backup.details.BackupDownloadDetailsViewModel;
-import org.wordpress.android.ui.jetpack.backup.progress.BackupDownloadProgressViewModel;
+import org.wordpress.android.ui.jetpack.backup.download.BackupDownloadViewModel;
+import org.wordpress.android.ui.jetpack.backup.download.complete.BackupDownloadCompleteViewModel;
+import org.wordpress.android.ui.jetpack.backup.download.details.BackupDownloadDetailsViewModel;
+import org.wordpress.android.ui.jetpack.backup.download.progress.BackupDownloadProgressViewModel;
 import org.wordpress.android.ui.jetpack.scan.ScanViewModel;
 import org.wordpress.android.ui.main.MeViewModel;
 import org.wordpress.android.ui.mediapicker.MediaPickerViewModel;
