+++ WordPress/src/main/java/org/wordpress/android/ui/jetpack/backup/download/BackupDownloadListItemState.kt
@@ -1,8 +1,13 @@
-package org.wordpress.android.ui.jetpack.backup
+package org.wordpress.android.ui.jetpack.backup.download
 
 import androidx.annotation.DrawableRes
 import org.wordpress.android.R
 import org.wordpress.android.ui.jetpack.BackupAvailableItemsProvider.BackupAvailableItemType
+import org.wordpress.android.ui.jetpack.backup.download.BackupDownloadListItemState.ViewType.ACTION_BUTTON
+import org.wordpress.android.ui.jetpack.backup.download.BackupDownloadListItemState.ViewType.CHECKBOX
+import org.wordpress.android.ui.jetpack.backup.download.BackupDownloadListItemState.ViewType.DESCRIPTION
+import org.wordpress.android.ui.jetpack.backup.download.BackupDownloadListItemState.ViewType.HEADER
+import org.wordpress.android.ui.jetpack.backup.download.BackupDownloadListItemState.ViewType.SUBHEADER
 import org.wordpress.android.ui.utils.UiString
 import org.wordpress.android.ui.utils.UiString.UiStringRes
 import org.wordpress.android.ui.utils.UiString.UiStringResWithParams
@@ -11,7 +16,7 @@ import org.wordpress.android.ui.utils.UiString.UiStringResWithParams
 sealed class BackupDownloadListItemState(val type: ViewType) {
     open fun longId(): Long = hashCode().toLong()
 
-    sealed class HeaderState : BackupDownloadListItemState(ViewType.HEADER) {
+    sealed class HeaderState : BackupDownloadListItemState(HEADER) {
         abstract val text: UiString
 
         data class DetailsHeaderState(
@@ -19,7 +24,7 @@ sealed class BackupDownloadListItemState(val type: ViewType) {
         ) : HeaderState()
     }
 
-    sealed class SubHeaderState : BackupDownloadListItemState(ViewType.SUBHEADER) {
+    sealed class SubHeaderState : BackupDownloadListItemState(SUBHEADER) {
         abstract val text: UiString
 
         data class DetailsSubHeaderState(
@@ -27,7 +32,7 @@ sealed class BackupDownloadListItemState(val type: ViewType) {
         ) : SubHeaderState()
     }
 
-    sealed class DescriptionState : BackupDownloadListItemState(ViewType.DESCRIPTION) {
+    sealed class DescriptionState : BackupDownloadListItemState(DESCRIPTION) {
         abstract val text: UiString
 
         data class DetailsSubHeaderState(
@@ -42,7 +47,7 @@ sealed class BackupDownloadListItemState(val type: ViewType) {
         }
     }
 
-    sealed class IconState : BackupDownloadListItemState(ViewType.DESCRIPTION) {
+    sealed class IconState : BackupDownloadListItemState(DESCRIPTION) {
         abstract val icon: Int
         abstract val contentDescription: UiString
 
@@ -53,7 +58,7 @@ sealed class BackupDownloadListItemState(val type: ViewType) {
         ) : IconState()
     }
 
-    sealed class ActionButtonState : BackupDownloadListItemState(ViewType.ACTION_BUTTON) {
+    sealed class ActionButtonState : BackupDownloadListItemState(ACTION_BUTTON) {
         abstract val text: UiString
         abstract val contentDescription: UiString
 
@@ -70,7 +75,7 @@ sealed class BackupDownloadListItemState(val type: ViewType) {
         val label: UiString,
         val checked: Boolean = false,
         val onClick: (() -> Unit)
-    ) : BackupDownloadListItemState(ViewType.CHECKBOX)
+    ) : BackupDownloadListItemState(CHECKBOX)
 
     enum class ViewType(val id: Int) {
         IMAGE(0),
