+++ WordPress/src/main/java/org/wordpress/android/ui/jetpack/backup/download/complete/BackupDownloadCompleteViewModel.kt
@@ -1,9 +1,9 @@
-package org.wordpress.android.ui.jetpack.backup.progress
+package org.wordpress.android.ui.jetpack.backup.download.complete
 
 import androidx.lifecycle.ViewModel
 import javax.inject.Inject
 
-class BackupDownloadProgressViewModel @Inject constructor() : ViewModel() {
+class BackupDownloadCompleteViewModel @Inject constructor() : ViewModel() {
     private var isStarted = false
 
     fun start() {
