+++ WordPress/src/test/java/org/wordpress/android/ui/jetpack/backup/download/BackupDownloadProgressCheckerTest.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.jetpack.backup
+package org.wordpress.android.ui.jetpack.backup.download
 
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.inOrder
